+++ CHANGELOG.md
@@ -10,3 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - using JWT boilerplate w. handlebars
 - refactoring client-side JS
 - added a test for userSchema method of 'checkPassword'
+
+## [twilio/1.0] - 2017-04-23
+### Added
+- beta version of the twilio basic application working
+
+## [twilio/2.1] - 2017-04-23
+### Added
+- iphone for the html page